Changed rest from int to float

+++ graph/schema/exercise.js
@@ -4,7 +4,7 @@ const Exercise = `
     name: String
     reps: Int
     sets: Int
-    rest: Int
+    rest: Float
     tempo: String
   }
 `;
